helix/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
flask-migrate==4.0.4
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==1.3.5
sqlalchemy==2.0.20
pytest==7.4.0
gunicorn==21.2.0
eventlet==0.33.3